file rename + update imports (#752)
