+++ server/src/http_server/routes/data_warehouse/data/graph_routes.ts
@@ -16,7 +16,7 @@ import WebGLFunctions from './graph_functions/webgl_functions';
 // Utilities
 import Config from '../../../../services/config';
 import Result from '../../../../common_classes/result';
-import ContainerRepository from '../../../../data_access_layer/repositories/data_warehouse/ontology/container_respository';
+import ContainerRepository from '../../../../data_access_layer/repositories/data_warehouse/ontology/container_repository';
 import RedisGraphService from '../../../../services/cache/redis_graph_loader';
 import {parseISO} from 'date-fns';
