+++ server/src/tests/data_warehouse/ontology/containers/template_repository.spec.ts
@@ -4,7 +4,7 @@ import PostgresAdapter from '../../../../data_access_layer/mappers/db_adapters/p
 import Logger from '../../../../services/logger';
 import Container, { DataSourceTemplate } from '../../../../domain_objects/data_warehouse/ontology/container';
 import UserMapper from '../../../../data_access_layer/mappers/access_management/user_mapper';
-import ContainerRepository from '../../../../data_access_layer/repositories/data_warehouse/ontology/container_respository';
+import ContainerRepository from '../../../../data_access_layer/repositories/data_warehouse/ontology/container_repository';
 import {KeyPair, User} from '../../../../domain_objects/access_management/user';
 import KeyPairRepository from '../../../../data_access_layer/repositories/access_management/keypair_repository';
 import UserRepository from '../../../../data_access_layer/repositories/access_management/user_repository';
